tpm2: use first PCR algorithm bank supported by TPM as default
The default is `sha1`, which is not always supported (it is legacy and
optional for implementation). Make this more future-proof and use the first
bank with non-empty set of PCRs, which is returned from TPM by
`tpm2_getcap pcrs`.

The swtpm by default does not create sha1 bank, so this fixes usage with
swtpm.

Signed-off-by: Oldřich Jedlička <oldium.pro@gmail.com>